requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
cmake==3.27.7
efficientnet-pytorch==0.7.1
filelock==3.13.1
fsspec==2023.10.0
huggingface-hub==0.19.0
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
Jinja2==3.1.2
lit==17.0.4
MarkupSafe==2.1.3
mpmath==1.3.0
munch==4.0.0
networkx==3.1
numpy==1.23.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-contrib-python==4.8.0.76
packaging==23.2
pandas==2.0.3
Pillow==10.0.1
pretrainedmodels==0.7.4
psutil==5.9.6
pydicom==2.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.0
scipy==1.8.1
segmentation-models-pytorch==0.3.3
six==1.16.0
sympy==1.12
timm==0.9.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
huggingface_hub==0.19.0
python-dotenv